+++ src/Core/ApieCore.php
@@ -7,13 +7,10 @@ use W2w\Lib\Apie\Apie;
 use W2w\Lib\Apie\Core\Resources\ApiResources;
 use W2w\Lib\Apie\OpenApiSchema\OpenApiSchemaGenerator;
 use W2w\Lib\Apie\OpenApiSchema\OpenApiSpecGenerator;
-use W2w\Lib\Apie\OpenApiSchema\SchemaGenerator;
 use W2w\Lib\Apie\OpenApiSchema\SubActions\SubActionContainer;
 use W2w\Lib\Apie\OpenApiSchema\SubActions\SubActionFactory;
 use W2w\Lib\Apie\PluginInterfaces\ResourceLifeCycleInterface;
 use W2w\Lib\Apie\PluginInterfaces\SubActionsProviderInterface;
-use W2w\Lib\Apie\Plugins\Core\Normalizers\ApieObjectNormalizer;
-use W2w\Lib\Apie\Plugins\Core\Normalizers\ContextualNormalizer;
 
 /**
  * Used by Apie to create the general Apie classes which you are not supposed to override in a plugin.
@@ -53,7 +50,7 @@ class ApieCore
     private $identifierExtractor;
 
     /**
-     * @var SchemaGenerator|null
+     * @var OpenApiSchemaGenerator|null
      */
     private $schemaGenerator;
 
@@ -111,29 +108,17 @@ class ApieCore
     /**
      * Returns the service that generates the JSON schema of a class.
      *
-     * @return SchemaGenerator
+     * @return OpenApiSchemaGenerator
      */
-    public function getSchemaGenerator(): SchemaGenerator
+    public function getSchemaGenerator(): OpenApiSchemaGenerator
     {
         if (!$this->schemaGenerator) {
-            if (ContextualNormalizer::isNormalizerEnabled(ApieObjectNormalizer::class)) {
-                $this->schemaGenerator = new SchemaGenerator(
-                    $this->apie->getClassMetadataFactory(),
-                    $this->apie->getPropertyTypeExtractor(),
-                    $this->getClassResourceConverter(),
-                    $this->apie->getPropertyConverter(),
-                    $this->apie->getDynamicSchemaLogic()
-                );
-            } else {
-                $this->schemaGenerator = new OpenApiSchemaGenerator(
-                    $this->apie->getDynamicSchemaLogic(),
-                    $this->apie->getObjectAccess(),
-                    $this->apie->getClassMetadataFactory(),
-                    $this->apie->getPropertyTypeExtractor(),
-                    $this->getClassResourceConverter(),
-                    $this->apie->getPropertyConverter()
-                );
-            }
+            $this->schemaGenerator = new OpenApiSchemaGenerator(
+                $this->apie->getDynamicSchemaLogic(),
+                $this->apie->getObjectAccess(),
+                $this->apie->getClassMetadataFactory(),
+                $this->apie->getPropertyConverter()
+            );
             foreach ($this->apie->getDefinedStaticData() as $class => $schema) {
                 $this->schemaGenerator->defineSchemaForResource($class, $schema);
             }
