--- src/Exceptions/ApieException.php
@@ -1,12 +1,0 @@
-<?php
-
-
-namespace W2w\Lib\Apie\Exceptions;
-
-/**
- * Base class that is extended by all Apie classes.
- * @deprecated use \W2w\Lib\ApieObjectAccessNormalizer\Exceptions\ApieException
- */
-abstract class ApieException extends \W2w\Lib\ApieObjectAccessNormalizer\Exceptions\ApieException
-{
-}
