+++ src/Exceptions/BadConfigurationException.php
@@ -1,6 +1,8 @@
 <?php
 namespace W2w\Lib\Apie\Exceptions;
 
+use W2w\Lib\ApieObjectAccessNormalizer\Exceptions\ApieException;
+
 class BadConfigurationException extends ApieException
 {
     public function __construct(
