+++ src/Exceptions/InvalidClassTypeException.php
@@ -1,7 +1,8 @@
 <?php
 
-
 namespace W2w\Lib\Apie\Exceptions;
+
+use W2w\Lib\ApieObjectAccessNormalizer\Exceptions\ApieException;
 
 /**
  * Exception thrown when a class is expected of a certain interface.
