+++ src/Exceptions/InvalidPageLimitException.php
@@ -1,8 +1,8 @@
 <?php
 
-
 namespace W2w\Lib\Apie\Exceptions;
 
+use W2w\Lib\ApieObjectAccessNormalizer\Exceptions\ApieException;
 use W2w\Lib\ApieObjectAccessNormalizer\Exceptions\LocalizationableException;
 use W2w\Lib\ApieObjectAccessNormalizer\Exceptions\LocalizationInfo;
