--- src/Exceptions/NameNotFoundException.php
@@ -1,11 +1,0 @@
-<?php
-
-
-namespace W2w\Lib\Apie\Exceptions;
-
-/**
- * @deprecated use \W2w\Lib\ApieObjectAccessNormalizer\Exceptions\NameNotFoundException
- */
-class NameNotFoundException extends \W2w\Lib\ApieObjectAccessNormalizer\Exceptions\NameNotFoundException
-{
-}
