+++ src/Exceptions/NotAnApiePluginException.php
@@ -1,6 +1,5 @@
 <?php
 
-
 namespace W2w\Lib\Apie\Exceptions;
 
 class NotAnApiePluginException extends BadConfigurationException
