+++ src/Exceptions/PageIndexShouldNotBeNegativeException.php
@@ -1,6 +1,7 @@
 <?php
 namespace W2w\Lib\Apie\Exceptions;
 
+use W2w\Lib\ApieObjectAccessNormalizer\Exceptions\ApieException;
 use W2w\Lib\ApieObjectAccessNormalizer\Exceptions\LocalizationableException;
 use W2w\Lib\ApieObjectAccessNormalizer\Exceptions\LocalizationInfo;
