--- src/Exceptions/ValidationException.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace W2w\Lib\Apie\Exceptions;
-
-/**
- * @deprecated use \W2w\Lib\ApieObjectAccessNormalizer\Exceptions\ValidationException
- */
-class ValidationException extends \W2w\Lib\ApieObjectAccessNormalizer\Exceptions\ValidationException
-{
-}
