+++ src/Plugins/Core/Serializers/SymfonySerializerAdapter.php
@@ -8,9 +8,6 @@ use W2w\Lib\Apie\Events\DecodeEvent;
 use W2w\Lib\Apie\Interfaces\FormatRetrieverInterface;
 use W2w\Lib\Apie\Interfaces\ResourceSerializerInterface;
 use W2w\Lib\Apie\PluginInterfaces\ResourceLifeCycleInterface;
-use W2w\Lib\Apie\Plugins\Core\Normalizers\ApieObjectNormalizer;
-use W2w\Lib\Apie\Plugins\Core\Normalizers\ContextualNormalizer;
-use W2w\Lib\Apie\Plugins\Core\Normalizers\EvilReflectionPropertyNormalizer;
 use W2w\Lib\ApieObjectAccessNormalizer\ObjectAccess\ObjectAccess;
 use Zend\Diactoros\Response\TextResponse;
 
@@ -91,13 +88,7 @@ class SymfonySerializerAdapter implements ResourceSerializerInterface
     }
 
     /**
-     * Decodes a request body to primitives. Forwards compatible with Apie version 4.
-     *
-     * @TODO change in inheritDoc in version 4
-     *
-     * @param string $requestBody
-     * @param string $contentType
-     * @return string|int|bool|float|array
+     * {@inheritDoc}
      */
     public function decodeRequestBody(string $requestBody, string $contentType)
     {
@@ -153,24 +144,11 @@ class SymfonySerializerAdapter implements ResourceSerializerInterface
      */
     public function hydrateWithReflection(array $data, string $resourceClass)
     {
-        if (!ContextualNormalizer::isNormalizerEnabled(ApieObjectNormalizer::class)) {
-            return $this->serializer->denormalize(
-                $data,
-                $resourceClass,
-                null,
-                ['object_access' => new ObjectAccess(false, true)]
-            );
-        }
-        ContextualNormalizer::enableDenormalizer(EvilReflectionPropertyNormalizer::class);
-        try {
-            return $this->serializer->denormalize(
-                $data,
-                $resourceClass,
-                null,
-                ['disable_type_enforcement' => true]
-            );
-        } finally {
-            ContextualNormalizer::disableDenormalizer(EvilReflectionPropertyNormalizer::class);
-        }
+        return $this->serializer->denormalize(
+            $data,
+            $resourceClass,
+           null,
+           ['object_access' => new ObjectAccess(false, true)]
+        );
     }
 }
