+++ src/Plugins/PrimaryKey/Normalizers/ApiePrimaryKeyNormalizer.php
@@ -9,8 +9,6 @@ use W2w\Lib\Apie\Core\ApiResourceMetadataFactory;
 use W2w\Lib\Apie\Core\ClassResourceConverter;
 use W2w\Lib\Apie\Core\IdentifierExtractor;
 use W2w\Lib\Apie\Core\Resources\ApiResourcesInterface;
-use W2w\Lib\Apie\Plugins\Core\Normalizers\ApieObjectNormalizer;
-use W2w\Lib\Apie\Plugins\Core\Normalizers\ContextualNormalizer;
 use W2w\Lib\Apie\Plugins\PrimaryKey\ValueObjects\PrimaryKeyReference;
 
 /**
@@ -63,7 +61,7 @@ class ApiePrimaryKeyNormalizer implements ContextAwareNormalizerInterface, Seria
      */
     public function supportsNormalization($data, $format = null, array $context = [])
     {
-        if (ContextualNormalizer::isNormalizerEnabled(ApieObjectNormalizer::class) || empty($context['object_hierarchy']) || !empty($context['disable_pk_normalize'])) {
+        if (empty($context['object_hierarchy']) || !empty($context['disable_pk_normalize'])) {
             return false;
         }
         foreach ($this->apiResources->getApiResources() as $apiResource) {
