+++ src/Plugins/PrimaryKey/Schema/ApiResourceLinkSchemaBuilder.php
@@ -6,10 +6,8 @@ namespace W2w\Lib\Apie\Plugins\PrimaryKey\Schema;
 
 use erasys\OpenApi\Spec\v3\Schema;
 use W2w\Lib\Apie\Core\ClassResourceConverter;
-use W2w\Lib\Apie\OpenApiSchema\SchemaGenerator;
+use W2w\Lib\Apie\OpenApiSchema\OpenApiSchemaGenerator;
 use W2w\Lib\Apie\PluginInterfaces\DynamicSchemaInterface;
-use W2w\Lib\Apie\Plugins\Core\Normalizers\ApieObjectNormalizer;
-use W2w\Lib\Apie\Plugins\Core\Normalizers\ContextualNormalizer;
 
 class ApiResourceLinkSchemaBuilder implements DynamicSchemaInterface
 {
@@ -27,16 +25,16 @@ class ApiResourceLinkSchemaBuilder implements DynamicSchemaInterface
      * @param string $operation
      * @param array $groups
      * @param int $recursion
-     * @param SchemaGenerator $generator
+     * @param OpenApiSchemaGenerator $generator
      */
     public function __invoke(
         string $resourceClass,
         string $operation,
         array $groups,
         int $recursion,
-        SchemaGenerator $generator
+        OpenApiSchemaGenerator $generator
     ) {
-        if (!ContextualNormalizer::isNormalizerEnabled(ApieObjectNormalizer::class) && $recursion > 0 && $operation === 'get') {
+        if ($recursion > 0 && $operation === 'get') {
             return new Schema([
                 'type' => 'string',
                 'format' => 'path',
