+++ src/Plugins/ValueObject/Schema/ValueObjectSchemaBuilder.php
@@ -3,7 +3,7 @@
 
 namespace W2w\Lib\Apie\Plugins\ValueObject\Schema;
 
-use W2w\Lib\Apie\OpenApiSchema\SchemaGenerator;
+use W2w\Lib\Apie\OpenApiSchema\OpenApiSchemaGenerator;
 use W2w\Lib\Apie\PluginInterfaces\DynamicSchemaInterface;
 
 class ValueObjectSchemaBuilder implements DynamicSchemaInterface
@@ -13,7 +13,7 @@ class ValueObjectSchemaBuilder implements DynamicSchemaInterface
         string $operation,
         array $groups,
         int $recursion,
-        SchemaGenerator $generator
+        OpenApiSchemaGenerator $generator
     ) {
         return $resourceClass::toSchema();
     }
