+++ tests/Plugins/PrimaryKey/Normalizers/ApiePrimaryKeyNormalizerTest.php
@@ -5,6 +5,7 @@ namespace W2w\Test\Apie\Plugins\PrimaryKey\Normalizers;
 
 use Doctrine\Common\Annotations\AnnotationReader;
 use LogicException;
+use PHPUnit\Framework\TestCase;
 use Symfony\Component\PropertyAccess\PropertyAccess;
 use Symfony\Component\Serializer\NameConverter\CamelCaseToSnakeCaseNameConverter;
 use Symfony\Component\Serializer\Serializer;
@@ -18,10 +19,9 @@ use W2w\Lib\Apie\Plugins\PrimaryKey\Normalizers\PrimaryKeyReferenceNormalizer;
 use W2w\Lib\Apie\Plugins\PrimaryKey\ValueObjects\PrimaryKeyReference;
 use W2w\Lib\Apie\Plugins\Uuid\Normalizers\UuidNormalizer;
 use W2w\Lib\ApieObjectAccessNormalizer\Normalizers\ApieObjectAccessNormalizer;
-use W2w\Test\Apie\ForwardsCompatibleTestCase;
 use W2w\Test\Apie\OpenApiSchema\Data\RecursiveObjectWithId;
 
-class ApiePrimaryKeyNormalizerTest extends ForwardsCompatibleTestCase
+class ApiePrimaryKeyNormalizerTest extends TestCase
 {
     /**
      * @var Serializer
