+++ tests/Plugins/PrimaryKey/PrimaryKeyPluginTest.php
@@ -4,13 +4,13 @@
 namespace W2w\Test\Apie\Plugins\PrimaryKey;
 
 use erasys\OpenApi\Spec\v3\Schema;
+use PHPUnit\Framework\TestCase;
 use W2w\Lib\Apie\DefaultApie;
 use W2w\Lib\Apie\Plugins\PrimaryKey\PrimaryKeyPlugin;
 use W2w\Lib\Apie\Plugins\StaticConfig\StaticResourcesPlugin;
-use W2w\Test\Apie\ForwardsCompatibleTestCase;
 use W2w\Test\Apie\OpenApiSchema\Data\RecursiveObjectWithId;
 
-class PrimaryKeyPluginTest extends ForwardsCompatibleTestCase
+class PrimaryKeyPluginTest extends TestCase
 {
     public function testSchemaIsProperlyCreated()
     {
